requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary  # or psycopg2, but psycopg2-binary is often easier to install
plotly
numpy
pygam
streamlit-aggrid